requirements.txt
django==5.1.4
djangorestframework==3.15.2
drf_yasg==1.21.8
psycopg==3.2.3
dj-rest-auth==7.0.0
django-allauth==65.3.1
requests==2.32.3